Horrible hack: kinda sorta support template literals
Currently string literals are represented as a `String` in Rust,
with no compile-time checking of the values. As long as this is the
case, template literals can be "implemented" by just pretending
they are `String`s as well.

This is horrible, but it doesn't *decrease* type safety, and allows
importing type definitions using template literals.

While we're at it, let's also flatten unions of strings into a single
`String`. There's no value gained by generating an enum with a single
`String(String)` variant.